refactor moving app_router to "common" directory
