+++ lib/common/router/app_router.dart
@@ -1,8 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 
-import '../features/authentication/presentation/screens/login_screen.dart';
-import '../features/authentication/presentation/screens/register_screen.dart';
-import '../features/on_boarding/presentation/screens/on_boarding_screen.dart';
+import '../../features/authentication/presentation/screens/login_screen.dart';
+import '../../features/authentication/presentation/screens/register_screen.dart';
+import '../../features/on_boarding/presentation/screens/on_boarding_screen.dart';
 
 part 'app_router.gr.dart';
