+++ test/common/router/app_router_test.dart
@@ -1,5 +1,5 @@
+import 'package:eatsy/common/router/app_router.dart';
 import 'package:eatsy/features/on_boarding/presentation/screens/on_boarding_screen.dart';
-import 'package:eatsy/utils/app_router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:responsive_builder/responsive_builder.dart';
